refactor: a common and a user api submodule
