+++ cloud-order-service/http/user.http
@@ -10,4 +10,4 @@ GET http://localhost:18082/actuator/loggers
 POST http://localhost:18082/actuator/loggers/com.frank.order.controller.IndexController
 Content-Type: application/json
 
-{"configuredLevel":"WARN"}
+{"configuredLevel":"DEBUG"}
