+++ cloud-order-service/src/main/java/com/frank/order/OrderApplication.java
@@ -1,5 +1,6 @@
 package com.frank.order;
 
+import com.frank.user.api.FeignUserService;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
@@ -9,7 +10,9 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
         DataSourceAutoConfiguration.class
 })
 @EnableDiscoveryClient
-@EnableFeignClients
+@EnableFeignClients(basePackageClasses = {
+        FeignUserService.class
+})
 public class OrderApplication {
 
         public static void main(String[] args) {
