+++ cloud-user-api/src/main/java/com/frank/user/api/FeignUserService.java
@@ -1,6 +1,6 @@
-package com.frank.order.service;
+package com.frank.user.api;
 
-import com.frank.user.model.dto.UserDTO;
+import com.frank.common.model.dto.UserDTO;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
