+++ pom.xml
@@ -14,7 +14,7 @@
         <module>cloud-item-service</module>
         <module>cloud-order-service</module>
         <module>cloud-demo-common</module>
-        <module>cloud-demo-api</module>
+        <module>cloud-user-api</module>
     </modules>
     <packaging>pom</packaging>
